+++ src/components/HeadUp/styles.css
