+++ src/index.js
@@ -1,13 +1,6 @@
 require('smoothscroll-polyfill').polyfill();
 
-import HeadUp from './containers/head-up';
-import Dashboard from './containers/dashboard';
-import Cell from './components/cell';
-import Menu from './components/menu';
-
-export {
-  HeadUp,
-  Dashboard,
-  Cell,
-  Menu
-};
+export {default as HeadUp} from './containers/HeadUp';
+export {default as Dashboard} from './containers/Dashboard';
+export {default as Cell} from './containers/Cell';
+export {default as Menu} from './components/Menu';
